Changed RateTeacher store to use Rating component. Tag validation works now. Some refactoring.
