+++ RateMyP/RateMyP.WebApp/ClientApp/src/store/index.ts
@@ -7,8 +7,7 @@ import * as CourseTeachers from './Course/CourseTeachers';
 import * as CourseTeacherActivities from './Course/CourseTeacherActivities';
 import * as Ratings from './Ratings';
 import * as Tags from './Tags';
-import * as Rate from './Rate/Rate';
-import * as RateTeacherActivites from './Rate/RateTeacherActivities'
+import * as Rate from './Rate/RateTeacher';
 import * as RateCourses from './Rate/RateCourses'
 
 // The top-level state object
@@ -24,9 +23,8 @@ export interface ApplicationState {
 
     ratings: Ratings.RatingsState | undefined;
     tags: Tags.TagsState | undefined;
-    rate: Rate.RateState | undefined;
-    rateTeacherActivites: RateTeacherActivites.RateTeacherActivitesState | undefined;
-    rateCourses: RateCourses.RateCoursesState | undefined;
+    rate: Rate.RateTeacherState | undefined; // rename to reateTeacher
+    rateCourses: RateCourses.RateCoursesState | undefined; // rename to RateCourse
 }
 
 // Whenever an action is dispatched, Redux will update each top-level application state property using
@@ -44,7 +42,6 @@ export const reducers = {
 
     ratings: Ratings.reducer,
     tags: Tags.reducer,
-    rateTeacherActivites: RateTeacherActivites.reducer,
     rate: Rate.reducer,
     rateCourses: RateCourses.reducer
 };
